chore(refactor-codebase): refactor routes folder into the server folder
- move the routes folder into server folder

[Delivers #161779044]
