+++ app.js
@@ -9,7 +9,7 @@ import session from 'express-session';
 import swaggerUi from 'swagger-ui-express';
 import swaggerDocument from './swagger.json';
 
-import routes from './routes';
+import routes from './server/routes';
 
 dotenv.config();
