+++ package.json
@@ -63,6 +63,7 @@
   },
   "devDependencies": {
     "@babel/node": "^7.0.0",
+    "@babel/register": "^7.0.0",
     "chai": "^4.2.0",
     "chai-http": "^4.2.0",
     "coveralls": "^3.0.2",
